Separate namespace for InjectionProcessor tests
