+++ AspectInjector.CompileTimeTests/InjectionProcessor/InjectionProcessorBaseClassTests.cs
@@ -1,18 +1,17 @@
 using System;
 using System.Linq;
 using AspectInjector.Broker;
-using AspectInjector.BuildTask.Processors.ModuleProcessors;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace AspectInjector.CompileTimeTests
+namespace AspectInjector.CompileTimeTests.InjectionProcessor
 {
     [TestClass]
-    public class InjectionProcessorBaseClassTests : InjectionProcessorTestBase
+    public class InjectionProcessorBaseClassTests : AInjectionProcessorTest
     {
         [TestMethod]
         public void Finds_Ctor_AspectContexts_BaseClass()
         {
-            var contexts = InjectionProcessor.GetAspectContexts(Module).ToArray();
+            var contexts = BuildTask.Processors.ModuleProcessors.InjectionProcessor.GetAspectContexts(Module).ToArray();
 
             Assert.AreEqual(1, contexts.Count(c => c.TargetName == ".ctor" && c.TargetTypeContext.TypeDefinition.Name == "TestClass"));
         }
@@ -20,7 +19,7 @@ namespace AspectInjector.CompileTimeTests
         [TestMethod]
         public void Finds_Method_AspectContexts_BaseClass()
         {
-            var contexts = InjectionProcessor.GetAspectContexts(Module).ToArray();
+            var contexts = BuildTask.Processors.ModuleProcessors.InjectionProcessor.GetAspectContexts(Module).ToArray();
             
             Assert.AreEqual(1, contexts.Count(c => c.TargetName == "VirtualMethodInBase" && c.TargetTypeContext.TypeDefinition.Name == "TestClass"));
             Assert.AreEqual(1, contexts.Count(c => c.TargetName == "VirtualMethodInBase" && c.TargetTypeContext.TypeDefinition.Name == "TestBaseClass"));
@@ -32,7 +31,7 @@ namespace AspectInjector.CompileTimeTests
         [TestMethod]
         public void Finds_Property_AspectContexts_BaseClass()
         {
-            var contexts = InjectionProcessor.GetAspectContexts(Module).ToArray();
+            var contexts = BuildTask.Processors.ModuleProcessors.InjectionProcessor.GetAspectContexts(Module).ToArray();
 
             Assert.AreEqual(2, contexts.Count(c => c.TargetName == "TestProperty"));
         }
@@ -40,7 +39,7 @@ namespace AspectInjector.CompileTimeTests
         [TestMethod]
         public void Finds_Event_AspectContexts_BaseClass()
         {
-            var contexts = InjectionProcessor.GetAspectContexts(Module).ToArray();
+            var contexts = BuildTask.Processors.ModuleProcessors.InjectionProcessor.GetAspectContexts(Module).ToArray();
 
             Assert.AreEqual(2, contexts.Count(c => c.TargetName == "TestEvent"));
         }
