+++ AspectInjector.CompileTimeTests/InjectionProcessor/InjectionProcessorInterfaceMemberTests.cs
@@ -1,18 +1,17 @@
 using System;
 using System.Linq;
 using AspectInjector.Broker;
-using AspectInjector.BuildTask.Processors.ModuleProcessors;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace AspectInjector.CompileTimeTests
+namespace AspectInjector.CompileTimeTests.InjectionProcessor
 {
     [TestClass]
-    public class InjectionProcessorInterfaceMemberTests : InjectionProcessorTestBase
+    public class InjectionProcessorInterfaceMemberTests : AInjectionProcessorTest
     {
         [TestMethod]
         public void Finds_Method_AspectContexts_InterfaceMember()
         {
-            var contexts = InjectionProcessor.GetAspectContexts(Module).ToArray();
+            var contexts = BuildTask.Processors.ModuleProcessors.InjectionProcessor.GetAspectContexts(Module).ToArray();
 
             Assert.AreEqual(1, contexts.Count(c => c.TargetName == "MethodInInterface"));
         }
@@ -20,7 +19,7 @@ namespace AspectInjector.CompileTimeTests
         [TestMethod]
         public void Finds_Property_AspectContexts_InterfaceMember()
         {
-            var contexts = InjectionProcessor.GetAspectContexts(Module).ToArray();
+            var contexts = BuildTask.Processors.ModuleProcessors.InjectionProcessor.GetAspectContexts(Module).ToArray();
 
             Assert.AreEqual(1, contexts.Count(c => c.TargetName == "PropertyInInterface"));
         }
@@ -28,7 +27,7 @@ namespace AspectInjector.CompileTimeTests
         [TestMethod]
         public void Finds_Event_AspectContexts_InterfaceMember()
         {
-            var contexts = InjectionProcessor.GetAspectContexts(Module).ToArray();
+            var contexts = BuildTask.Processors.ModuleProcessors.InjectionProcessor.GetAspectContexts(Module).ToArray();
 
             Assert.AreEqual(2, contexts.Count(c => c.TargetName == "EventInInterface"));
         }
